Raise MSVC warning level from /W3 to /W4
This patch increases the warning level when using the MSVC compiler from
`/W3` to `/W4` and fixes the issues found. Four warnings introduced by
`/W4` are disabled, all related to variable name shadowing, as they
overly prescriptive to valid code.

+++ source/adapters/cuda/program.cpp
@@ -11,6 +11,8 @@
 #include "program.hpp"
 #include "ur_util.hpp"
 
+#include <array>
+
 bool getMaxRegistersJitOptionValue(const std::string &BuildOptions,
                                    unsigned int &Value) {
   using namespace std::string_view_literals;
@@ -137,8 +139,8 @@ ur_result_t ur_program_handle_t_::buildProgram(const char *BuildOptions) {
   }
 
   UR_CHECK_ERROR(cuModuleLoadDataEx(&Module, static_cast<const void *>(Binary),
-                                    Options.size(), Options.data(),
-                                    OptionVals.data()));
+                                    static_cast<unsigned int>(Options.size()),
+                                    Options.data(), OptionVals.data()));
 
   BuildStatus = UR_PROGRAM_BUILD_STATUS_SUCCESS;
